Rearrange schema loader integration tests
