+++ schema-loader/src/integration-test/java/com/scalar/db/server/SchemaLoaderWithScalarDbServer.java
@@ -1,11 +1,8 @@
-package com.scalar.db.schemaloader.server;
+package com.scalar.db.server;
 
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
-import com.scalar.db.server.ScalarDbServer;
-import com.scalar.db.server.ServerConfig;
-import com.scalar.db.server.ServerEnv;
 import java.io.IOException;
 import org.junit.jupiter.api.AfterAll;
