+++ schema-loader/src/integration-test/java/com/scalar/db/storage/cosmos/CosmosSchemaLoaderIntegrationTest.java
@@ -1,8 +1,7 @@
-package com.scalar.db.schemaloader.cosmos;
+package com.scalar.db.storage.cosmos;
 
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
-import com.scalar.db.storage.cosmos.CosmosEnv;
 import java.util.Optional;
 
 public class CosmosSchemaLoaderIntegrationTest extends SchemaLoaderIntegrationTestBase {
