+++ schema-loader/src/integration-test/java/com/scalar/db/storage/dynamo/DynamoSchemaLoaderWithStorageSpecificArgsIntegrationTest.java
@@ -1,36 +1,39 @@
-package com.scalar.db.schemaloader.jdbc;
+package com.scalar.db.storage.dynamo;
 
 import com.google.common.collect.ImmutableList;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
-import com.scalar.db.storage.jdbc.JdbcEnv;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import org.junit.jupiter.api.Disabled;
 
-public class JdbcSchemaLoaderWithStorageSpecificArgsIntegrationTest
+public class DynamoSchemaLoaderWithStorageSpecificArgsIntegrationTest
     extends SchemaLoaderIntegrationTestBase {
 
   @Override
   protected DatabaseConfig getDatabaseConfig() {
-    return JdbcEnv.getJdbcConfig();
+    return DynamoEnv.getDynamoConfig();
   }
 
   @Override
   protected List<String> getCommandArgsForCreationWithCoordinatorTable(
       String configFile, String schemaFile) throws IOException {
-    DatabaseConfig config = new DatabaseConfig(new File(configFile));
+    DynamoConfig config = new DynamoConfig(new File(configFile));
     return ImmutableList.of(
-        "--jdbc",
-        "-j",
+        "--dynamo",
+        "--region",
         config.getContactPoints().get(0),
         "--schema-file",
         schemaFile,
         "-u",
         config.getUsername().get(),
         "-p",
-        config.getPassword().get());
+        config.getPassword().get(),
+        "--endpoint-override",
+        config.getEndpointOverride().get(),
+        "--no-scaling",
+        "--no-backup");
   }
 
   @Disabled
