+++ schema-loader/src/integration-test/java/com/scalar/db/storage/jdbc/JdbcSchemaLoaderWithStorageSpecificArgsIntegrationTest.java
@@ -1,41 +1,35 @@
-package com.scalar.db.schemaloader.dynamo;
+package com.scalar.db.storage.jdbc;
 
 import com.google.common.collect.ImmutableList;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
-import com.scalar.db.storage.dynamo.DynamoConfig;
-import com.scalar.db.storage.dynamo.DynamoEnv;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import org.junit.jupiter.api.Disabled;
 
-public class DynamoSchemaLoaderWithStorageSpecificArgsIntegrationTest
+public class JdbcSchemaLoaderWithStorageSpecificArgsIntegrationTest
     extends SchemaLoaderIntegrationTestBase {
 
   @Override
   protected DatabaseConfig getDatabaseConfig() {
-    return DynamoEnv.getDynamoConfig();
+    return JdbcEnv.getJdbcConfig();
   }
 
   @Override
   protected List<String> getCommandArgsForCreationWithCoordinatorTable(
       String configFile, String schemaFile) throws IOException {
-    DynamoConfig config = new DynamoConfig(new File(configFile));
+    DatabaseConfig config = new DatabaseConfig(new File(configFile));
     return ImmutableList.of(
-        "--dynamo",
-        "--region",
+        "--jdbc",
+        "-j",
         config.getContactPoints().get(0),
         "--schema-file",
         schemaFile,
         "-u",
         config.getUsername().get(),
         "-p",
-        config.getPassword().get(),
-        "--endpoint-override",
-        config.getEndpointOverride().get(),
-        "--no-scaling",
-        "--no-backup");
+        config.getPassword().get());
   }
 
   @Disabled
