+++ schema-loader/src/integration-test/java/com/scalar/db/storage/multistorage/MultiStorageSchemaLoaderIntegrationTest.java
@@ -1,9 +1,7 @@
-package com.scalar.db.schemaloader.multistorage;
+package com.scalar.db.storage.multistorage;
 
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.schemaloader.SchemaLoaderIntegrationTestBase;
-import com.scalar.db.storage.multistorage.MultiStorageConfig;
-import com.scalar.db.storage.multistorage.MultiStorageEnv;
 import com.scalar.db.transaction.consensuscommit.Coordinator;
 import java.util.Properties;
